requirements.txt
markdown==3.7
tomli; python_version < '3.11'
PyYAML==6.0.2
